Precision Upgrade 2: Electric Boogaloo

+++ src/core/SampleBuffer.cpp
@@ -1025,7 +1025,7 @@ void SampleBuffer::visualize(
 	const float ySpace = h * 0.5f;
 	const int nbFrames = focusOnRange ? toFrame - fromFrame : m_frames;
 
-	const float fpp = std::max(1, nbFrames / w);
+	const double fpp = std::max(1, nbFrames / w);
 	// There are 2 possibilities: Either nbFrames is bigger than
 	// the width, so we will have width points, or nbFrames is
 	// smaller than the width (fpp = 1) and we will have nbFrames
@@ -1050,7 +1050,7 @@ void SampleBuffer::visualize(
 		? fromFrame + visibleFrames - 1
 		: visibleFrames - 1;
 
-	for (float frame = first; frame <= last && frame <= lastVisibleFrame; frame += fpp)
+	for (double frame = first; frame <= last && frame <= lastVisibleFrame; frame += fpp)
 	{
 		float maxData = -1;
 		float minData = 1;
